# openapi/lostandfound.yaml
openapi: 3.0.1
info:
  title: Lost & Found API
  description: API for managing lost and found items.
  version: 1.0.0
servers:
  - url: http://localhost:9191
    description: Local server

paths:
  /user/lost:
    get:
      summary: Get all lost items
      tags:
        - Lost Items
      responses:
        '200':
          description: A list of all lost items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LostItemDto'
        '500':
          description: Server error

  /user/claim:
    post:
      summary: Claim a lost item
      tags:
        - Claims
      requestBody:
        description: Details of the claimed item
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimItemRequest'
      responses:
        '200':
          description: Item claimed successfully
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad request
        '500':
          description: Server error

  /get-claims:
    get:
      summary: Get all claimed items
      tags:
        - Claims
      responses:
        '200':
          description: A list of all claimed items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClaimedItemDto'
        '500':
          description: Server error

components:
  schemas:
    LostItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the lost item
        itemName:
          type: string
          description: Name of the lost item
        quantity:
          type: integer
          description: Quantity of the lost item
        place:
          type: string
          description: Place where the item was lost
      required:
        - itemName
        - quantity
        - place

    ClaimItemRequest:
      type: object
      properties:
        lostItemId:
          type: integer
          format: int64
          description: ID of the lost item
        userId:
          type: integer
          format: int64
          description: ID of the user claiming the item
        quantity:
          type: integer
          description: Quantity of the lost item
      required:
        - userId
        - lostItem

    ClaimedItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the claimed item
        userId:
          type: integer
          format: int64
          description: ID of the user who claimed the item
        itemName:
          type: string
          description: Name of the lost item
        quantity:
          type: integer
          description: Quantity of the lost item
        place:
          type: string
          description: Place where the item was lost
      required:
        - userId
        - itemName
        - quantity
        - place